fix(Methods): change process increasing and decreasing stock and salesRate
Now the Menu class manages state of stock instead of the JpaMenuRepository class as it concentrates on CRUD process.
Even though the fields in entity should be mapped with data in database, stock and salesRate data in the entity and those in the database have not been matched in the meantime.
Changed to having same values.